Projectile, boomerang
Added projectile parent class for all moving projectiles that go in a straight line. Bullets will be set to use this parent class in the future.

Added the boomerang. Press E to throw it, it will come back when it hits an emu. Will make it return after a certain point if it does not hit in the future.